# [fts]: Utilizing an optimized index for fts