--- a/migrations/0001_add_advocates_search_vector.sql
+++ b/migrations/0001_add_advocates_search_vector.sql
@@ -1,0 +1,10 @@
+CREATE INDEX "idx_advocates_computed_fts" ON "advocates" USING gin (
+          to_tsvector('english', 
+            "first_name" || ' ' || 
+            "last_name" || ' ' || 
+            "city" || ' ' || 
+            "degree" || ' ' ||
+            COALESCE("specialties"::text, '') || ' ' ||
+            COALESCE("years_of_experience"::text, '')
+          )
+        );