--- a/package.json
+++ b/package.json
@@ -30,7 +30,7 @@
     "@types/node": "^22.18.5",
     "@types/react": "^19.1.13",
     "@types/react-dom": "^19.1.9",
-    "drizzle-kit": "^0.23.0",
+    "drizzle-kit": "^0.31.4",
     "drizzle-seed": "^0.3.1",
     "eslint": "^9.35.0",
     "eslint-config-next": "^15.5.3",