# What Changed?

For now we will maintain the global full-text search, but as-is it is really inefficient in how it runs.

PostgreSQL has a feature that would make the same search we are already doing efficient and scalable. So as a first pass to improving the search, we add in this index and utilize it during search.

So we first update our schema file to add this new index. A GIN index (Generalized Inverted Index) on the full-text search vector allows PostgreSQL to perform extremely fast text searches across multiple columns (name, city, degree, specialties, experience) in
milliseconds instead of slow sequential scans. It works by creating an inverted index structure that maps each word to the records containing it, enabling efficient queries like
"find all advocates with 'family lawyer boston'" without having to scan every row.

Then we just need to utilize the vector search functions to ensure we utilize that index!

Here is a query plan for the old `ILIKE` setup:
```
EXPLAIN ANALYSE SELECT *
FROM advocates
WHERE
    first_name ILIKE '%pain%' OR
    last_name ILIKE '%pain%' OR
    city ILIKE '%pain%' OR
    degree ILIKE '%pain%' OR
    COALESCE(specialties::text, '') ILIKE '%pain%' OR
    COALESCE(years_of_experience::text, '') ILIKE '%pain%'
;

Seq Scan on advocates  (cost=0.00..2.75 rows=1 width=114) (actual time=0.446..0.851 rows=3 loops=1)
"  Filter: (((first_name)::text ~~* '%pain%'::text) OR ((last_name)::text ~~* '%pain%'::text) OR ((city)::text ~~* '%pain%'::text) OR ((degree)::text ~~* '%pain%'::text) OR (COALESCE((specialties)::text, ''::text) ~~* '%pain%'::text) OR (COALESCE((years_of_experience)::text, ''::text) ~~* '%pain%'::text))"
  Rows Removed by Filter: 47
Planning Time: 1.203 ms
Execution Time: 1.087 ms
```
Kind of hard to read, but the important call out is the sequential scan. The table is currently small but once this grows this will become really detrimental to performance. Anytime the database has to read the entire table row by row, you know you wont scale well.

Here is the same query run with the new index and vector queries:
```
EXPLAIN ANALYSE SELECT *
FROM advocates
WHERE to_tsvector('english', first_name || ' ' || last_name || ' ' || city || ' ' || degree || ' ' || COALESCE(specialties::text, '') || ' ' || COALESCE(years_of_experience::text, ''))
    @@ plainto_tsquery('english', 'pain')
;

Bitmap Heap Scan on advocates  (cost=8.55..12.85 rows=1 width=114) (actual time=0.069..0.071 rows=3 loops=1)
"  Recheck Cond: (to_tsvector('english'::regconfig, (((((((((((first_name)::text || ' '::text) || (last_name)::text) || ' '::text) || (city)::text) || ' '::text) || (degree)::text) || ' '::text) || COALESCE((specialties)::text, ''::text)) || ' '::text) || COALESCE((years_of_experience)::text, ''::text))) @@ '''pain'''::tsquery)"
  Heap Blocks: exact=1
  ->  Bitmap Index Scan on idx_advocates_computed_fts  (cost=0.00..8.55 rows=1 width=0) (actual time=0.016..0.016 rows=3 loops=1)
"        Index Cond: (to_tsvector('english'::regconfig, (((((((((((first_name)::text || ' '::text) || (last_name)::text) || ' '::text) || (city)::text) || ' '::text) || (degree)::text) || ' '::text) || COALESCE((specialties)::text, ''::text)) || ' '::text) || COALESCE((years_of_experience)::text, ''::text))) @@ '''pain'''::tsquery)"
Planning Time: 0.259 ms
Execution Time: 0.156 ms
``` 
The big thing to notice here is the Index Scan. It will first use the index to find the matching rows, then in the Heap Scan will fetch the found rows.

## Links
- https://orm.drizzle.team/docs/guides/postgresql-full-text-search
- https://www.postgresql.org/docs/current/textsearch-controls.html
- https://www.postgresql.org/docs/current/gin.html

## Screenshot/Video

## Related PRs


### Reminders

- Link to the related ticket
- Tests added or updated
- Add/Update related documentation